+++ ogn/parser/telnet_parser.py
@@ -23,11 +23,18 @@ def parse(telnet_data):
                 'turn_rate': float(telnet_data[101:107]),
                 'magic_number': int(telnet_data[114:116]),
                 'gps_status': telnet_data[117:122],
-                'frequency_offset': float(telnet_data[123:129]),
-                'signal_quality': float(telnet_data[132:137]),
-                'error_count': float(telnet_data[139:142]),
-                'distance': float(telnet_data[143:150]),
-                'bearing': float(telnet_data[152:158]),
-                'phi': float(telnet_data[161:167])}
+                'channel': int(telnet_data[124:126]),
+                'flarm_timeslot': telnet_data[126] == 'f',
+                'ogn_timeslot': telnet_data[127] == 'o',
+                'frequency_offset': float(telnet_data[128:134]),
+                'decode_quality': float(telnet_data[137:142]),
+                'signal_quality': float(telnet_data[143:147]),
+                'demodulator_type': int(telnet_data[150:151]),
+                'error_count': float(telnet_data[151:154]),
+                'distance': float(telnet_data[155:162]),
+                'bearing': float(telnet_data[164:170]),
+                'phi': float(telnet_data[173:179]),
+                'multichannel': telnet_data[183] == '+'}
+
     except Exception:
         raise ParseError