+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # CHANGELOG
+## unreleased
+- parser: fixed telnet parser
+
 ## 0.9.4: - 2019-06-10
 - parser: Added support for OGINREACH (Garmin inReach) beacons
 - parser: Added support for OGFLYM (Flymaster) beacons